+++ src/index.ts
@@ -4,6 +4,10 @@ export * from './types/interfaces';
 export * from './types/enums';
 export * from './conversations/conversation-update';
 export * from 'softphone-vendor-headsets';
+export * from './sessions/softphone-session-handler';
+export * from './sessions/video-session-handler';
+export * from './sessions/screen-recording-session-handler';
+export * from './sessions/screen-share-session-handler';
 
 import * as utils from './utils';
 import * as mediaUtils from './media/media-utils';
@@ -17,4 +21,5 @@ export { SdkError } from './utils';
 
 /* make sure to export the default */
 import GenesysCloudWebrtSdk from './client';
+
 export default GenesysCloudWebrtSdk;