+++ test/test-utils.ts
@@ -19,6 +19,10 @@ export class SimpleMockSdk extends EventEmitter {
     jest.spyOn(SdkMedia.prototype, 'initialize' as any).mockReturnValue(null);
     this.media = new SdkMedia(this as any as GenesysCloudWebrtcSdk);
     this._http = new HttpClient();
+
+    const messenger = new EventEmitter();
+    (messenger as any).broadcastMessage = jest.fn();
+    (this._streamingConnection as any).messenger = messenger;
   }
 
   media: SdkMedia;
@@ -43,11 +47,17 @@ export class SimpleMockSdk extends EventEmitter {
     log: jest.fn(),
     info: jest.fn(),
     warn: jest.fn(),
-    error: jest.fn()
+    error: jest.fn(),
+    config: {
+      appName: 'MockSdkApp'
+    }
   };
   _streamingConnection = {
     config: {
       jid: 'userjid@orgspan.com'
+    },
+    activeStanzaInstance: {
+      jid: 'myjid@genesys.com/resource'
     },
     notifications: {
       subscribe: jest.fn(),
@@ -64,7 +74,8 @@ export class SimpleMockSdk extends EventEmitter {
     },
     _webrtcSessions: {
       refreshIceServers: jest.fn(),
-      getSessionManager: jest.fn()
+      getSessionManager: jest.fn(),
+      proxyNRStat: jest.fn()
     }
   };
   sessionManager = {