+++ test/unit/sessions/video-session-handler.test.ts
@@ -8,9 +8,9 @@ import BaseSessionHandler from '../../../src/sessions/base-session-handler';
 import { SessionTypes, CommunicationStates, SdkErrorTypes } from '../../../src/types/enums';
 import * as mediaUtils from '../../../src/media/media-utils';
 import * as utils from '../../../src/utils';
-import { IParticipantsUpdate, IExtendedMediaSession, IConversationParticipant, VideoMediaSession } from '../../../src/types/interfaces';
+import { IParticipantsUpdate, IExtendedMediaSession, IConversationParticipant, VideoMediaSession, MemberStatusMessage } from '../../../src/types/interfaces';
 import VideoSessionHandler, { IMediaChangeEvent } from '../../../src/sessions/video-session-handler';
-import { ConversationUpdate, IMemberStatusMessage } from '../../../src/';
+import { ConversationUpdate, GenesysDataChannelMessageParams } from '../../../src/';
 import { JsonRpcMessage } from 'genesys-cloud-streaming-client';
 
 let handler: VideoSessionHandler;
@@ -1493,7 +1493,7 @@ describe('handleDataChannelMessage', () => {
   it('should call handleMemberStatusMessage', () => {
     const spy = handler.handleMemberStatusMessage = jest.fn();
 
-    const message: IMemberStatusMessage = {
+    const message: JsonRpcMessage = {
       jsonrpc: 'v2',
       method: 'member.notify.status',
       params: {
@@ -1529,7 +1529,7 @@ describe('handleMemberStatusMessage', () => {
       emit: jest.fn()
     };
     const spy = session.emit = jest.fn();
-    const message: IMemberStatusMessage = {
+    const message: MemberStatusMessage = {
       jsonrpc: 'v2',
       method: 'member.notify.status',
       params: {