+++ changelog.md
@@ -3,7 +3,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v8.2.4...HEAD)
+# [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v9.0.0...HEAD)
+
+# [v9.0.0](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v8.2.4...v9.0.0)
+### BREAKING CHANGES
+* Only one sdk instance will be allowed to have headset controls at any given time. Headset functionality is now orchestrated and negotiated between other instances of the sdk based on various factors. As such, we have changed the DeviceConnectionStatus type in order to union those states with orchestration states.
+
+### Added
+* [PCM-2224](https://inindca.atlassian.net/browse/PCM-2224) - Added headset orchestration flow. Opt out with `config.disableHeadsetControlsOrchestration`, but this will be temporary and will be removed without notice.
+* [PCM-2250](https://inindca.atlassian.net/browse/PCM-2250) - Emit `sessionInterrupted` event when a session's `connectionState` becomes `interrupted`.
+* [PCM-2276](https://inindca.atlassian.net/browse/PCM-2276) - Added nr events for media actions and first alerting conversation update
 
 # [v8.2.4](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v8.2.3...v8.2.4)
 * [PCM-2238](https://inindca.atlassian.net/browse/PCM-2238) - Bumped client logger to help with a fix for media helper